Adds optional host parameter to users.
All calls that involve calling a user by name now
also allow for the host to be specified optionally.

Similarly, all calls that respond with a user also include
the host, defaulting to '%', MySQL's shorthand for 'anywhere'.

Hosts with dots in the name are escaped to avoid a problem
with the way our routing parses URLs.

Change-Id: Idc5d514a7d862a723469ca8b49f1c51ae07f741b