Allow for compatable auths between new and old CLI
The new CLI allows for Rax auth, but alas it does not work the same way
as the old CLI did. This means that any scripts using both have to have
their auth URLs switched.

The fix for this is to include a new auth type for the old CLI tool
that adds "/tokens" to the auth URL the way the new CLI does.

Fixes bug 1253746

Change-Id: I5705d87033cc42b9dfa136ca13e4e5ab83bd666d