Changed Trove CLI list headers to match Nova CLI
Added ability to pass in header labels to print_list.
If not passed in, header lables are automatically created
from the fields by replacing '_' with spaces and capitalizing
each word.  IDs are also capitalized.  All field data is
now left-aligned, except numerics which are right-aligned.

Also changed print_dict so that all values are left-aligned.

Change-Id: I3b69dc7f92c496594cf37832f4ff98f1abdf52dd
Closes-Bug: #1272700