Fallback to flavor's str_id when id is None
When a nova flavor's id is a string and not an int, Trove will
return None for the flavor's id and instead expect clients
to use the str_id field.  This updates the shell mask this
from the user and always present the relevant value as id, and
allow specifying either str_id or id for flavor-show.

find_resource() was updated to deal with ints and strings, not just
UUIDs and ints.

Relies on changes to Trove @ https://review.openstack.org/#/c/115811

Related-bug: #1333852

Change-Id: If08430b07b7e8b6a1737f3e71dba6a471de63794