default endpoint_type to 'publicURL'
When trying to create a troveclient.v1.client.Client without specifying
endpoint_type explicitly it will raise an AuthorizationFailure exception.
This was because troveclient.v1.client.Client defaulted endpoint_type=None
which overrides the troveclient.client.HTTPClient default value of 'publicURL'
to None. Fix was to default the endpoint_type to 'publicURL' in the
troveclient.v1.client.Client __init__ method.

Also updated the documentation and README.rst for creating a
trove client object because it was out of date and incorrect.

Change-Id: I64dfbaa4eefc73d01468470dbe9eba5e209ff157
Closes-Bug: 1409982