Fixed output of cluster-create in CLI
When running cluster-create, the CLI prints the instances created as
part of the output.  This causes the results to scroll off the screen
and makes it very difficult to see what happened.

The output of cluster-create has been consolidated with that of
cluster-show, in much the same way as is done with create and show
for single instances.

Change-Id: Id4eecc236b6c1ec43b47714ac5eabd8323a5b996
Closes-Bug: #1563504