Remove times.dbm file for each tox run
If you run the py27 tox tests then attempt to run the py34 tests,
the following error will occur:

    running testr
    running=OS_STDOUT_CAPTURE=${OS_STDOUT_CAPTURE:-1} \
    OS_STDERR_CAPTURE=${OS_STDERR_CAPTURE:-1} \
    OS_TEST_TIMEOUT=${OS_TEST_TIMEOUT:-60} \
    ${PYTHON:-python} -m subunit.run discover -t ./ ./ --list
    db type could not be determined
    error: testr failed (3)

This happens because the apparently the times.dbm file generated
by py27 is not compatible with py34. The converse is not true though -
if you run py34 first, py27 will work.

Both these bugs have been filed against it, but I see no resolution
https://bugs.launchpad.net/testrepository/+bug/1212909
https://bugs.launchpad.net/testrepository/+bug/1229445

The work-around is to delete the file, which is what is now done
before running every tox test. Once the bug has been fixed, the
work-around can be removed.

Change-Id: Icf5d9efc65ffda8acdcc6f43b37ef03c84b3fd43
Closes-Bug: #1565928