Locality support for clusters
In order to allow clusters to be all on the same hypervisor
(affinity) or all on different hypervisors (anti-affinity) a new
argument (locality) needed to be added to the Trove cluster-create API.

This changeset addresses the Trove client part of this feature.
A --locality flag is now available on the 'cluster-create' command and
is passed to the server for processing.

Partially implements: blueprint replication-cluster-locality
Change-Id: Ie46cfa69ffe6a64760aca38c495563f5724bd0d0