Add support for module ordering on apply
Client side of being able to specify 'priority' modules plus a
way to rank the order in which modules are applied.  Two new
attributes 'priority_apply' and 'apply_order' have been added
to the python API and are supported in the CLI by means of the
module-create and module-update commands.

Partially Implements: blueprint module-management-ordering

Change-Id: I2b871c7d1eab634174396bd0905329f3a66755f6