Fix module-list
Under a keystone V3 scenario, the CLI command for
module-list will fail as it tried to determine whether
to display the 'visible' and 'tenant' fields (which only
make sense to show to an admin).

Keystone V3 stores the role information (used to find
out if the user has admin access) in a different
place than V2.  This fix uses a keystoneauth1 property
to have the value determined correctly internally,
so trove doesn't have to worry about it.

The logic was also moved to the utils module to
faciliate reuse.

Closes-Bug: #1622019
Change-Id: I6dbc3660b507017f85d06bde2903f4d2334fea35