Throw correct error on creation of size 0
When trying to create an instance of size 0 an error is thrown
stating that the size was not specified, this is not correct
because I did specify it. The problem is that args.size evaluates
to False because size is 0.

This patch checks for size being zero instead and throws the correct
error message.

Change-Id: If00e83ccd919429dcfac7621dbbd9daa49f8b416