Enable Keystone v3 support for compat client
Enable Keystone v3 auth support in the compat client.
This is needed for the Trove integration and scenario tests which
use the compat client. The Keystone v2 APIs have been removed in
the Queens release and this stop-gap fix of putting Keystone v3
support in the compat client is needed to fix the Trove gate.

Change-Id: I60ea2188443d8905c03c24607703e69c6490420b