Fix print_list label name in troveclient/utils.py
The output of list make nice labels from fields. Fisrt, replace
underscore with space. Then, replace "id" with "ID". Finally,
capitalize each word. This works fine for "*_id"(for exmaple,
"flavor_id" will be replaced to "Flavor ID"), but when a word
has a "id" inside, it will be wrong. For example, "cidr"
will be replaced to "CIDr", actually it should be "Cidr".

This change fix the problem, only "id" after a underscore will be
replaced to "ID".

Change-Id: I8e357c946b21dc6f6c6c79d78a45f23b440a57da