Fix error with  configuration-parameter-show
According bug description, there will be an error when
use  configuration-parameter-show without version_id
and datastore. To fix this problem, I add an judgement
for use the cmd without parameter datstore and
datastore_version not like an uuid.

Now in such a case, the cmd will has an exception info.

Change-Id: I01fd56bad80f389da369be90dff5c89b05e6a8dc
Closes-Bug: #1363065