add lower-constraints job
Create a tox environment for running the unit tests against the lower
bounds of the dependencies.

Create a lower-constraints.txt to be used to enforce the lower bounds
in those tests.

Add openstack-tox-lower-constraints job to the zuul configuration.

See http://lists.openstack.org/pipermail/openstack-dev/2018-March/128352.html
for more details.

About PrettyTable requirement:
The minimum version of PrettyTable is not correct in requirements.txt,
which should be 0.7.2 according to the unittests(before 0.7.2, if the
table is empty, i.e. no rows in the table, PrettyTable won't print the
headers and borders, this is changed during 0.7.2)

Co-Authored-By: Zhao Chao <zhaochao1984@gmail.com>
Change-Id: Ia9cadcbc2bc9aa2864fe47040dfe7be8ae286b98
Depends-On: https://review.openstack.org/555034
Signed-off-by: Doug Hellmann <doug@doughellmann.com>