Correctly mocking builtin open in test_v1_shell
There's now a seperate module 'builtins' from python-future(this package
is now installed by stestr, and stestr in required by oslotest-3.6.0),
so under Python 2.x 'builtins' is also imported as a module and
'bultins.open' is mocked instead of '__builtin__.open'. And according to
the Mock document, we should always patch the object where it's looked
up, so only necessary mocking is made in this patch.

Change-Id: I57ea61cefd3b01b19475b0fa7e16f716eaaa14e6
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>